Fix sometimes glitchy dirt bucket placement
Took 27 minutes

+++ src/main/java/com/ferreusveritas/dynamictrees/items/DirtBucket.java
@@ -17,6 +17,7 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.level.ClipContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult;
 
@@ -79,8 +80,9 @@ public class DirtBucket extends Item {
     }
 
     public boolean tryPlaceContainedDirt(@Nullable Player player, Level world, BlockPos posIn) {
-        if (world.getBlockState(posIn).getMaterial().isReplaceable()) {
-            if (!world.isClientSide) {
+        BlockState blockState = world.getBlockState(posIn);
+        if (blockState.getMaterial().isReplaceable()) {
+            if (!world.isClientSide && !blockState.isAir()) {
                 world.destroyBlock(posIn, true);
             }
